+++ src/utils/musicConstants.js
@@ -1,0 +1,11 @@
+export const notes = ['C', 'C#', 'D', 'Eb', 'E', 'F', 'F#', 'G', 'Ab', 'A', 'Bb', 'B'];
+
+export const modes = {
+    Ionian:     [0, 2, 4, 5, 7, 9, 11],
+    Dorian:     [0, 2, 3, 5, 7, 9, 10],
+    Phrygian:   [0, 1, 3, 5, 7, 8, 10],
+    Lydian:     [0, 2, 4, 6, 7, 9, 11],
+    Mixolydian: [0, 2, 4, 5, 7, 9, 10],
+    Aeolian:    [0, 2, 3, 5, 7, 8, 10],
+    Locrian:    [0, 1, 3, 5, 6, 8, 10]
+};
